LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/contain-intrinsic-size/contain-intrinsic-size-029-expected.txt


PASS .test 1
FAIL .test 2 assert_equals:
<button class="test cis-height" data-expected-client-width="0" data-expected-client-height="50"></button>
clientHeight expected 50 but got 0
FAIL .test 3 assert_equals:
<button class="test cis-width" data-expected-client-width="100" data-expected-client-height="0"></button>
clientWidth expected 100 but got 0
FAIL .test 4 assert_equals:
<button class="test cis-both" data-expected-client-width="100" data-expected-client-height="50"></button>
clientWidth expected 100 but got 0
PASS .test 5
FAIL .test 6 assert_equals:
<button class="test cis-height min-size" data-expected-client-width="20" data-expected-client-height="50"></button>
clientHeight expected 50 but got 5
FAIL .test 7 assert_equals:
<button class="test cis-width min-size" data-expected-client-width="100" data-expected-client-height="5"></button>
clientWidth expected 100 but got 20
FAIL .test 8 assert_equals:
<button class="test cis-both min-size" data-expected-client-width="100" data-expected-client-height="50"></button>
clientWidth expected 100 but got 20
